refactor warp and masque. extract account provisioning into generator.go
Move the load-from-cache-or-create-and-save logic and createConfig out of the outbound/endpoint startHandler into a new generator.go in each package, and expose GenerateConfig(recreate) for reuse. run behavior is unchanged.
