+++ protocol/warp/endpoint.go
@@ -2,16 +2,13 @@ package warp
 
 import (
 	"context"
-	"encoding/json"
 	"math/rand"
 	"net"
 	"net/netip"
 	"strings"
-	"time"
 
 	"github.com/sagernet/sing-box/adapter"
 	"github.com/sagernet/sing-box/adapter/endpoint"
-	"github.com/sagernet/sing-box/common/cloudflare"
 	C "github.com/sagernet/sing-box/constant"
 	"github.com/sagernet/sing-box/log"
 	"github.com/sagernet/sing-box/option"
@@ -21,8 +18,6 @@ import (
 	"github.com/sagernet/sing/common/json/badoption"
 	M "github.com/sagernet/sing/common/metadata"
 	N "github.com/sagernet/sing/common/network"
-	"github.com/sagernet/sing/service"
-	"golang.zx2c4.com/wireguard/wgctrl/wgtypes"
 )
 
 func RegisterEndpoint(registry *endpoint.Registry) {
@@ -55,36 +50,10 @@ func NewEndpoint(ctx context.Context, router adapter.Router, logger log.ContextL
 	}
 	endpoint.startHandler = func() {
 		defer close(endpoint.await)
-		cacheFile := service.FromContext[adapter.CacheFile](ctx)
-		var config *Config
-		var err error
-		if !options.Profile.Recreate && cacheFile != nil && cacheFile.StoreWARPConfig() {
-			savedProfile := cacheFile.LoadWARPConfig(tag)
-			if savedProfile != nil {
-				if err = json.Unmarshal(savedProfile.Content, &config); err != nil {
-					logger.ErrorContext(ctx, err)
-					return
-				}
-			}
-		}
-		if config == nil {
-			config, err = endpoint.createConfig()
-			if err != nil {
-				logger.ErrorContext(ctx, err)
-				return
-			}
-			if cacheFile != nil && cacheFile.StoreWARPConfig() {
-				content, err := json.Marshal(config)
-				if err != nil {
-					logger.ErrorContext(ctx, err)
-					return
-				}
-				cacheFile.SaveWARPConfig(tag, &adapter.SavedBinary{
-					LastUpdated: time.Now(),
-					Content:     content,
-					LastEtag:    "",
-				})
-			}
+		config, err := endpoint.provisionConfig(options.Profile.Recreate)
+		if err != nil {
+			logger.ErrorContext(ctx, err)
+			return
 		}
 		peer := config.Peers[0]
 		hostParts := strings.Split(peer.Endpoint.Host, ":")
@@ -179,48 +148,4 @@ func (w *Endpoint) isEndpointInitialized(ctx context.Context) error {
 		return E.New("endpoint not initialized")
 	}
 	return nil
-}
-
-func (w *Endpoint) createConfig() (*Config, error) {
-	opts := make([]cloudflare.CloudflareApiOption, 0, 1)
-	if w.options.Profile.Detour != "" {
-		detour, ok := service.FromContext[adapter.OutboundManager](w.ctx).Outbound(w.options.Profile.Detour)
-		if !ok {
-			return nil, E.New("outbound detour not found: ", w.options.Profile.Detour)
-		}
-		opts = append(opts, cloudflare.WithDialContext(func(ctx context.Context, network, addr string) (net.Conn, error) {
-			return detour.DialContext(ctx, network, M.ParseSocksaddr(addr))
-		}))
-	}
-	var privateKey wgtypes.Key
-	var err error
-	if w.options.Profile.PrivateKey != "" {
-		privateKey, err = wgtypes.ParseKey(w.options.Profile.PrivateKey)
-		if err != nil {
-			return nil, err
-		}
-	} else {
-		privateKey, err = wgtypes.GeneratePrivateKey()
-		if err != nil {
-			return nil, err
-		}
-	}
-	api := cloudflare.NewCloudflareApi(opts...)
-	var profile *cloudflare.CloudflareProfile
-	if w.options.Profile.AuthToken != "" && w.options.Profile.ID != "" {
-		profile, err = api.GetProfile(w.ctx, w.options.Profile.AuthToken, w.options.Profile.ID)
-		if err != nil {
-			return nil, err
-		}
-	} else {
-		profile, err = api.CreateProfile(w.ctx, privateKey.PublicKey().String())
-		if err != nil {
-			return nil, err
-		}
-	}
-	return &Config{
-		PrivateKey: privateKey.String(),
-		Interface:  profile.Config.Interface,
-		Peers:      profile.Config.Peers,
-	}, nil
 }
